Renames module, adds oauth to count file
